refactor(fleet-wiki): clean-code pass over the cowork surfaces
- Absorb the 22-line runtime.ts into service.ts (its only consumer);
  the ./cowork public surface is unchanged via the barrel.
- Drop dead exports and vocabulary: unused CoworkState / CoworkAnnotation
  aliases, the never-emitted "draft" event type, the consumer-less
  client fetchCoworkTranscript with its transcript route, and the
  client-side CoworkTranscriptTurnDto re-export (transcript stays a
  server-internal feed for one-shot history).
