+++ packages/fleet-wiki/src/cowork/index.ts
@@ -2,6 +2,5 @@
 // 호스트(Fleet Console 등)는 HTTP/SSE 어댑터만 소유하고 이 엔진을 소비한다.
 export * from "./types.js";
 export { CoworkStore } from "./store.js";
-export { CoworkService } from "./service.js";
+export { CoworkService, createCoworkMcpRuntime } from "./service.js";
 export type { CoworkAgentClient, CoworkConnectOptions, CoworkConnector } from "./service.js";
-export { createCoworkMcpRuntime } from "./runtime.js";
