+++ runtime/fleet-console/core/host/codex/api-types.ts
@@ -3,7 +3,7 @@
 // 이 파일에 Node 타입(NodeJS.* 등)을 참조하면 안 된다.
 
 // Cowork DTO의 단일 출처는 fleet-wiki cowork 서브패키지다(type-only라 브라우저 번들 무영향).
-export type { CoworkAnnotationDto, CoworkEventDto, CoworkSessionDto, CoworkTranscriptTurnDto } from "@dotobokuri/fleet-wiki/cowork";
+export type { CoworkAnnotationDto, CoworkEventDto, CoworkSessionDto } from "@dotobokuri/fleet-wiki/cowork";
 /** 콘솔 options 라우트 계약 — CLI/모델 목록은 호스트가 core-unified-agent 레지스트리에서 채운다. */
 export interface CoworkOptionsResponse { clis: readonly string[]; models: readonly string[]; efforts: readonly string[]; defaultModel?: string; defaultEffort?: string; }
